+++ Randomizer.OutputTests/Tests/Integer/IntegerOutputTest.cs
@@ -5,13 +5,13 @@ namespace Randomizer.OutputTests.Tests.Integer
 {
     public class IntegerOutputTest : OutputTestBase
     {
-        protected IRandomInteger randomLong;
+        protected IRandomInteger randomInteger;
 
-        public IntegerOutputTest(IRandomInteger randomLong, ILogger fileLogger)
+        public IntegerOutputTest(IRandomInteger randomInteger, ILogger fileLogger)
         {
-            Validator.ValidateNull(randomLong);
+            Validator.ValidateNull(randomInteger);
             Validator.ValidateNull(fileLogger);
-            this.randomLong = randomLong;
+            this.randomInteger = randomInteger;
             FileLogger = fileLogger;
         }
